DataDog-friendly OpenTelemetry metadata
The OpenTelemetry API/SDK, since 1.1.0, automatically
adds opentelemetry-related log metadata to processes that have a trace context set.

This is extremely useful as it allows all opentelemetry-instrumented code to
have opentelemetry log metadata automatically, without the need for each
instrumenting library to set the metadata explicitly whenever setting the
current span.

Unfortunately, the metadata set by the SDK doesn't play nicely with DataDog:
it doesn't use the keys that datadog expects to make logs-APM correlation work,
and additionally it's set a charlists values which are serialized to unsightly
list of integers by prima_ex_logger.

This commit introduces a new option to prima_logger_ex so transform the
OpenTelemetry metadata to some different representations that work better with
Prima's infrastructure.

The new default is the :datadog format, which will transform the SDK
opentelemetry metadata into datadog-friendly equivalents.
Check the README for other options.