Don't break if OpenTelemetry SDK metadata has unexpected values/types
Adds code to handle unexpected metadata values in the metadata
keys "reserved" by the OpenTelemetry SDK; and a test to verify
that the logger doesn't break in such cases.
This could happen if the OpenTelemetry SDK were to upgrade the format
of its logger metadata, or if the application using PrimaExLogger
is overwriting those metadata keys with their own arbitrary values.